Added a world builder, collisions, the core entities and a basic input tracker

+++ core/src/com/bressio/rendezvous/helpers/PhysicsManager.java
@@ -5,5 +5,5 @@ import com.badlogic.gdx.math.Vector2;
 public class PhysicsManager {
 
     public static final Vector2 gravity = new Vector2(0, 0);
-    public static final int PPM = 100;
+    public static final float PPM = 100;
 }
